rewrite the installer process. Impossible to not make it work now.

+++ setup.sh
@@ -37,7 +37,22 @@
 
 case $1 in
     install)
+    cd ~
+    if [ $USER == "root" ] then
         clear
+        echo "DO NOT USE ROOT"
+        echo "  _____           _        _ _                   _                _           _    ";
+        echo " |_   _|         | |      | | |                 | |              | |         | |   ";
+        echo "   | |  _ __  ___| |_ __ _| | | ___ _ __    __ _| |__   ___  _ __| |_ ___  __| |   ";
+        echo "   | | | '_ \/ __| __/ _\` | | |/ _ \ '__|  / _\` | '_ \ / _ \| '__| __/ _ \/ _\` |";
+        echo "  _| |_| | | \__ \ || (_| | | |  __/ |    | (_| | |_) | (_) | |  | ||  __/ (_| |   ";
+        echo " |_____|_| |_|___/\__\__,_|_|_|\___|_|     \__,_|_.__/ \___/|_|   \__\___|\__,_|   ";
+        echo "                                                                                   ";
+        echo "                                                                                   ";
+        echo "DO NOT USE ROOT"
+        exit
+    fi
+    clear
         echo " __      _______  _        _____           _        _ _            ";
         echo " \ \    / /  __ \| |      |_   _|         | |      | | |           ";
         echo "  \ \  / /| |__) | |        | |  _ __  ___| |_ __ _| | | ___ _ __  ";
@@ -60,14 +75,12 @@ case $1 in
             echo "                                                                                   ";
             ;;
             *)
-            cd && sudo chown -R $USER:$USER ~/* ; sudo chown -R $USER:$USER ~/.*
-            read -p "Do you want to update the system? [Y/n]: " updated
-            case $updated in
-                n|N|no|No|NO)
-                echo "Moving on..."
-                PATH=$PATH:~/.local/bin
-                ;;
-                *)
+            cd
+            sudo chown -R $USER:$USER ~/* ; sudo chown -R $USER:$USER ~/.*
+            if [ -d "byob" ]; then
+                rm -rf ~/byob
+            fi
+            if ! sudo apt update | grep -woc "All packages are up to date" >; then
                 # Don't do any actions before user agrees to the terms.
                 sudo apt update && sudo apt upgrade -y && sudo apt full-upgrade -y
                 sudo apt autoremove -y
@@ -82,13 +95,13 @@ case $1 in
                 echo "                                   ";
                 read -p "Your system must now reboot (Press [enter])"
                 sudo reboot now
-                ;;
-            esac
+            fi
+            PATH=$PATH:~/.local/bin
             echo "Applying pre-perms to service files"
-                sudo cp ~/vrl-package/byob-server.service /etc/systemd/system/ > /dev/null
-                sudo cp ~/vrl-package/byob /usr/bin/ > /dev/null
-                sudo chown root:root /etc/systemd/system/byob-server.service > /dev/null
-            sleep .5
+            sudo cp ~/vrl-package/byob-server.service /etc/systemd/system/ \
+            ; sudo cp ~/vrl-package/byob /usr/bin/ \
+            ; sudo chown root:root /etc/systemd/system/byob-server.service \
+            ; clear > /dev/null
             echo "  _    _           _       _   _                   ";
             echo " | |  | |         | |     | | (_)                  ";
             echo " | |  | |_ __   __| | __ _| |_ _ _ __   __ _       ";
@@ -97,85 +110,66 @@ case $1 in
             echo "  \____/| .__/ \__,_|\__,_|\__|_|_| |_|\__, (_|_|_)";
             echo "        | |                             __/ |      ";
             echo "        |_|                            |___/       ";
+            echo "Fetching fresh meat..."
             sleep .5
-            echo "Installing dependencies..."
-                echo "Fetching fresh meat..."
-                    sudo apt install docker.io \
-                    git gcc cmake make upx-ucl \
-                    build-essential zlib1g-dev \
-                    neofetch htop avahi-daemon -qy  > /dev/null
-                echo "Picking python repos..."
-                    sudo apt install python3 python3-pip python3-opencv python3-wheel \
-                    python3-setuptools python3-dev python3-distutils python3-venv -qy  > /dev/null
-                echo "Doing magic..."
-                    sudo systemctl start avahi-daemon  > /dev/null
-                    sudo systemctl enable avahi-daemon  > /dev/null
-                    sudo systemctl start docker  > /dev/null
-                    sudo systemctl enable docker > /dev/null
-                echo "Cloning vrlnx/BYOB files..."
-                    sleep .4
-                    git -C ~/ clone https://github.com/vrlnx/byob.git > /dev/null
-                echo "Installing BYOB dependencies..."
-                    cd ~/byob/byob
-                    sleep .5
-                    python3 setup.py  > /dev/null
-                    pip3 install -r requirements.txt  > /dev/null
-                    pip3 install colorama  > /dev/null
-                    pip3 install pyinstaller==3.6  > /dev/null
-                    pip3 install numpy==1.17.3 > /dev/null
-                    pip3 install requests > /dev/null
-                    pip3 install flask > /dev/null
-                    pip3 install flask_wtf > /dev/null
-                    pip3 install flask_mail > /dev/null
-                    pip3 install flask-bcrypt > /dev/null
-                    pip3 install flask-login > /dev/null
-                    pip3 install flask-sqlalchemy > /dev/null
-                    cd
-                echo "Set permissions..."
-                    sleep 2
-                    chmod +x ~/vrl-package/uninstaller.sh
-                    chmod +x ~/vrl-package/start-byob.sh
-                echo "Transfer Ownership..."
-                    echo "Current user: $USER"
-                    sudo usermod -aG docker $USER > /dev/null
-                    sudo chown -R $USER:$USER ~/byob > /dev/null
-                    read -p "Are you ready to use the system? (Enter to continue)"
-            clear
-            echo "  ______     ______  ____    _____           _        _ _          _   ";
-            echo " |  _ \ \   / / __ \|  _ \  |_   _|         | |      | | |        | |  ";
-            echo " | |_) \ \_/ / |  | | |_) |   | |  _ __  ___| |_ __ _| | | ___  __| |  ";
-            echo " |  _ < \   /| |  | |  _ <    | | | '_ \/ __| __/ _\` | | |/ _ \/ _\` |";
-            echo " | |_) | | | | |__| | |_) |  _| |_| | | \__ \ || (_| | | |  __/ (_| |  ";
-            echo " |____/  |_|  \____/|____/  |_____|_| |_|___/\__\__,_|_|_|\___|\__,_|  ";
-            echo "                                                                       ";
-            echo "                                                                       ";
+            echo "Doing magic..."
             echo " "
-            echo "Enabled start-byob.sh"
-            echo "Your website will open on ""$HOSTNAME"".local:5000"" when server is started."
-            cd ~/vrl-package
-            read -p "Do you want to start BYOB - Open Source web-server? [Y/n]" webServer
-            case $webServer in
-                n|N)
-                echo "You start BYOB - Open Source Webserver by running"
-                echo "'./start-byob.sh' inside 'vrl-package' folder"
-                ;;
-                *)
-                ./start-byob.sh
-                ;;
-            esac
-            ;;
+            echo "Sit back and enjoy a drink, this may take a while..."
+            echo "Do not cancel... (If not installed after 1 hour, then there is trouble...)"
+                sudo apt install docker.io git gcc cmake make upx-ucl build-essential zlib1g-dev \
+                neofetch htop avahi-daemon \
+                python3 python3-pip python3-opencv python3-wheel python3-setuptools \
+                python3-dev python3-distutils python3-venv -y > /dev/null
+                ; sudo systemctl start avahi-daemon \
+                ; sudo systemctl enable avahi-daemon \
+                ; sudo systemctl start docker \
+                ; sudo systemctl enable docker \
+                ; git -C ~/ clone https://github.com/vrlnx/byob.git \
+                ; cd ~/byob/byob \
+                ; python3 setup.py \
+                ; pip3 install -r requirements.txt \
+                ; pip3 install colorama \
+                ; pip3 install pyinstaller==3.6 \
+                ; pip3 install numpy==1.17.3 \
+                ; pip3 install requests \
+                ; pip3 install flask \
+                ; pip3 install flask_wtf \
+                ; pip3 install flask_mail \
+                ; pip3 install flask-bcrypt \
+                ; pip3 install flask-login \
+                ; pip3 install flask-sqlalchemy \
+                ; pip3 install wtforms \
+                ; cd \
+                ; chmod +x ~/vrl-package/uninstaller.sh \
+                ; chmod +x ~/vrl-package/start-byob.sh \
+                ; sudo usermod -aG docker $USER \
+                ; sudo chown -R $USER:$USER ~/byob > /dev/null
+        clear
+        chmod -x ~/vrl-package/setup.sh
+        echo "  ______     ______  ____    _____           _        _ _          _   ";
+        echo " |  _ \ \   / / __ \|  _ \  |_   _|         | |      | | |        | |  ";
+        echo " | |_) \ \_/ / |  | | |_) |   | |  _ __  ___| |_ __ _| | | ___  __| |  ";
+        echo " |  _ < \   /| |  | |  _ <    | | | '_ \/ __| __/ _\` | | |/ _ \/ _\` |";
+        echo " | |_) | | | | |__| | |_) |  _| |_| | | \__ \ || (_| | | |  __/ (_| |  ";
+        echo " |____/  |_|  \____/|____/  |_____|_| |_|___/\__\__,_|_|_|\___|\__,_|  ";
+        echo "                                                                       ";
+        echo "                                                                       ";
+        echo " "
+        echo "Enabled start-byob.sh"
+        echo "#1 Type 'newgrp docker', hit enter"
+        echo "#2 Type './start-byob.sh', hit enter"
+        cd ~/vrl-package
         esac
         ;;
     help)
-        echo "DO NOT USE ROOT USER!"
+        clear
         echo "Commands around this setup:"
         echo "./setup.sh help"
         echo "./setup.sh install - Start installing"
         echo "./setup.sh rules - Show EULA"
         ;;
     rules)
-        echo "DO NOT USE ROOT USER!"
-        echo ""
+        clear
         echo "VRLinux, the owner, support team,"
         echo "or anyone else is not responsible for your use of this framework."
         echo ""
@@ -185,7 +179,9 @@ case $1 in
         echo "> use './setup.sh install' to start installing"
         ;; 
     *)
-    echo "DO NOT USE ROOT USER!"
-    echo "> Use './setup.sh rules' to read EULA"
+    clear
+    echo "> Welcome to VRL Installer"
+    echo "> This installer need you to read our rules."
+    echo "> Type './setup.sh rules' to read EULA"
     ;;
 esac
